+++ core/src/packet/packet_pw.cpp
@@ -1,0 +1,4 @@
+void AOClient::pktCharPassword(AreaData* area, int argc, QStringList argv, AOPacket packet)
+{
+    password = argv[0];
+}
