+++ core/src/packet/packet_rc.cpp
@@ -1,0 +1,4 @@
+void AOClient::pktRequestChars(AreaData* area, int argc, QStringList argv, AOPacket packet)
+{
+    sendPacket("SC", server->characters);
+}
